bashrc
This one contains a few functions tacked onto the Ubuntu 10.10 default cruft. These functions wrap ssh and sftp, and set the XTerm's title. This ensures that the title will always reflect where the terminal is connected to, even when connecting to a machine that doesn't have a PS1 with xterm escape codes to update it regularly.